requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
asyncio==3.4.3
dnspython==2.4.1
ffmpeg-python==0.2.0
GitPython==3.1.32
pillow
psutil==5.9.5
pyrogram
python-dotenv==1.0.0
requests==2.31.0
speedtest-cli==2.1.3
TgCrypto==1.2.5
motor
youtube-search-python
